+++ packages/v1/react-textarea/CHANGELOG.md
@@ -1,5 +1,23 @@
 # ui
 
+## 1.54.0-next.4
+
+### Patch Changes
+
+- Updated dependencies [af11b1e]
+  - @copilotkit/react-core@1.54.0-next.4
+  - @copilotkit/runtime-client-gql@1.54.0-next.4
+  - @copilotkit/shared@1.54.0-next.4
+
+## 1.54.0-next.3
+
+### Patch Changes
+
+- Updated dependencies [fa0d1cd]
+  - @copilotkit/react-core@1.54.0-next.3
+  - @copilotkit/runtime-client-gql@1.54.0-next.3
+  - @copilotkit/shared@1.54.0-next.3
+
 ## 1.53.1-next.2
 
 ### Patch Changes