+++ packages/v1/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkit/runtime-client-gql
 
+## 1.54.0-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.54.0-next.4
+
+## 1.54.0-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.54.0-next.3
+
 ## 1.53.1-next.2
 
 ### Patch Changes