+++ packages/v2/sqlite-runner/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkitnext/sqlite-runner
 
+## 1.54.0-next.4
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.54.0-next.4
+
+## 1.54.0-next.3
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.54.0-next.3
+
 ## 1.53.1-next.2
 
 ### Patch Changes